Fixed version to int

+++ teraserver/python/libtera/db/Base.py
@@ -15,7 +15,7 @@ class BaseModel:
     # cause problems with the mapper)
     __mapper_args__ = {
         'version_id_col': version_id,
-        'version_id_generator': lambda version: time.time()*1000  # uuid.uuid4().hex
+        'version_id_generator': lambda version: int(time.time()*1000)  # uuid.uuid4().hex
     }
 
     def to_json(self, ignore_fields=None):
